+++ src/classes/Solution.ts
@@ -23,6 +23,33 @@ export class Solution
             const lineText = textLine.text.trim();
             const lowerCase = lineText.toLowerCase();
 
+            if(start === undefined)
+            {
+                if(this.FileFormat === undefined
+                && lowerCase.startsWith("microsoft visual studio solution file, format version"))
+                {
+                    this.FileFormat = textLine;
+                }
+
+                if(this.VersionComment === undefined
+                && lowerCase.startsWith("# visual studio version"))
+                {
+                    this.VersionComment = textLine;
+                }
+
+                if(this.StudioVersion === undefined
+                && lowerCase.startsWith("visualstudioversion"))
+                {
+                    this.StudioVersion = textLine;
+                }
+
+                if(this.MinimumStudioVersion === undefined
+                && lowerCase.startsWith("minimumvisualstudioversion"))
+                {
+                    this.MinimumStudioVersion = textLine;
+                }
+            }
+
             if(lowerCase === "global")
             {
                 start = textLine.range.start;
@@ -40,4 +67,12 @@ export class Solution
     public Projects: Array<Project>;
 
     public Global?: Global;
+
+    public FileFormat: vscode.TextLine|undefined;
+
+    public VersionComment: vscode.TextLine|undefined;
+
+    public StudioVersion: vscode.TextLine|undefined;
+
+    public MinimumStudioVersion: vscode.TextLine|undefined;
 }